refactor: [#639] Tracker Checker: extract mod for HTTP checks
